Use build constraint for cloneTLSConfig
